+++ CHANGES.rst
@@ -18,6 +18,9 @@
   if it is only an instance of `BaseException` but not of `Exception`,
   such as e.g. a ``SystemExit`` or ``KeyboardInterupt`` exception.
 
+- Support abort hooks (symmetrically to commit hooks)
+  (`#77 <https://github.com/zopefoundation/transaction/issues/77>`_).
+
 
 2.4.0 (2018-10-23)
 ==================